requirements.txt
fastapi
uvicorn
langgraph
langchain-google-genai  # For Gemini Pro
langchain-openai         # For GPT-4o
motor                    # Asynchronous MongoDB driver
arxiv                    # ArXiv API wrapper
python-dotenv
pydantic
numpy
pandas
httpx                    # Async HTTP client
apscheduler              # Feature 14: Scheduled auto-runs
pdfplumber               # Feature 16: PDF text extraction
certifi                  # SSL certificate bundle
